Update for using numpy.1.24; require numpy.1.20 or later

+++ examples/subduction-3d/utils/slip_invert.py
@@ -170,7 +170,7 @@ class SlipInvert(Application):
         vertsF = s.create_dataset('geometry/vertices', data=self.faultVertCoords)
         timesF = s.create_dataset('time', data=timesF, maxshape=(None, 1, 1))
         topoF = s.create_dataset('viz/topology/cells', data=self.faultCells, dtype='d')
-        topoF.attrs['cell_dim'] = numpy.int32(cellDimF)
+        topoF.attrs['cell_dim'] = numpy.int64(cellDimF)
         slipVec = numpy.zeros((self.numPenaltyWeights, self.numFaultVerts, 3),
                               dtype=numpy.float64)
         slipAlongRake = numpy.zeros((self.numPenaltyWeights, self.numFaultVerts, 1),
@@ -183,7 +183,7 @@ class SlipInvert(Application):
         vertsD = d.create_dataset('geometry/vertices', data=self.dataCoords)
         timesD = d.create_dataset('time', data=timesD, maxshape=(None, 1, 1))
         topoD = d.create_dataset('viz/topology/cells', data=topolD, dtype='d')
-        topoD.attrs['cell_dim'] = numpy.int32(cellDimD)
+        topoD.attrs['cell_dim'] = numpy.int64(cellDimD)
 
         predictedDisp = numpy.zeros((self.numPenaltyWeights, self.numDataPoints, 3), dtype=numpy.float64)
 
@@ -283,7 +283,7 @@ class SlipInvert(Application):
         self.faultVertCoords = impulsesLl['geometry/vertices'][:]
         self.numFaultVerts = self.faultVertCoords.shape[0]
         self.faultCells = numpy.array(impulsesLl['viz/topology/cells'][:],
-                                      dtype=numpy.int)
+                                      dtype=numpy.int64)
         self.numFaultCells = self.faultCells.shape[0]
         llSlip = impulsesLl['vertex_fields/slip'][:,:, 0]
         llImpInds = numpy.nonzero(llSlip != 0.0)
